Move mbed-trace into platform
