+++ UNITTESTS/CMakeLists.txt
@@ -114,6 +114,7 @@ set(unittest-includes-base
   "${PROJECT_SOURCE_DIR}/../features"
   "${PROJECT_SOURCE_DIR}/../platform/include"
   "${PROJECT_SOURCE_DIR}/../platform/include/platform"
+  "${PROJECT_SOURCE_DIR}/../platform/mbed-trace/include"
   "${PROJECT_SOURCE_DIR}/../storage/filesystem/littlefs/include"
   "${PROJECT_SOURCE_DIR}/../storage/filesystem/fat/include"
   "${PROJECT_SOURCE_DIR}/../storage/blockdevice/include"
@@ -132,7 +133,6 @@ set(unittest-includes-base
   "${PROJECT_SOURCE_DIR}/../events/source"
   "${PROJECT_SOURCE_DIR}/../rtos/include"
   "${PROJECT_SOURCE_DIR}/../features/frameworks"
-  "${PROJECT_SOURCE_DIR}/../features/frameworks/mbed-trace"
   "${PROJECT_SOURCE_DIR}/../connectivity/libraries/nanostack-libservice"
   "${PROJECT_SOURCE_DIR}/../connectivity/libraries/nanostack-libservice/mbed-client-libservice"
   "${PROJECT_SOURCE_DIR}/../connectivity/netsocket/include"
