+++ UNITTESTS/empty_baseline/unittest.cmake
@@ -8,7 +8,7 @@ set(unittest-includes ${unittest-includes}
   ..
   ../connectivity/mbedtls/include/mbedtls/
   ../connectivity/mbedtls/platform/inc/
-  ../features/frameworks/mbed-trace/mbed-trace/
+  ../platform/mbed-trace/mbed-trace/
 )
 
 set(unittest-test-sources
