+++ platform/mbed-trace/.gitignore
