+++ platform/mbed-trace/.mbedignore
