+++ platform/mbed-trace/LICENSE
