+++ platform/mbed-trace/mbed-trace/mbed_trace.h
