+++ platform/mbed-trace/mbed_lib.json
