+++ platform/mbed-trace/source/CMakeLists.txt
