+++ platform/mbed-trace/source/mbed_trace.c
